pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "color-match-tools"
version = "5.6.1"
description = "Comprehensive color science library for accurate color space conversions, perceptual color distance metrics (Delta E), and color matching with CSS and 3D printing filament databases"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "David Terracino", email = "dterracino@gmail.com" }
]
keywords = ["color", "color-science", "delta-e", "color-conversion", "color-matching", "rgb", "lab", "lch", "hsl", "xyz", "ciede2000", "cie94", "cie76", "gamut", "srgb", "3d-printing", "filament"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

# No external dependencies - pure Python stdlib!
dependencies = []

[project.optional-dependencies]
# Enhanced fuzzy matching for validation module
fuzzy = [
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.12.0",  # Accelerates fuzzywuzzy
]

# Image processing features (requires Pillow and numpy)
image = [
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-image>=0.20.0",
    "pillow-heif>=0.14.0",  # HEIC format support (AVIF supported natively in Pillow 10+)
]

# Documentation generation
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

# Development tools
dev = [
    "pyright>=1.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

# Everything (all optional features)
all = [
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.12.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-image>=0.20.0",
    "pillow-heif>=0.14.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/dterracino/color_tools"
Repository = "https://github.com/dterracino/color_tools"
Issues = "https://github.com/dterracino/color_tools/issues"
Changelog = "https://github.com/dterracino/color_tools/blob/main/CHANGELOG.md"
Documentation = "https://dterracino.github.io/color_tools/"

[project.scripts]
color_tools = "color_tools.cli:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["color_tools", "color_tools.*"]
exclude = ["color_tools.data*", "tests*", "test_*"]

[tool.setuptools.package-data]
color_tools = ["data/*.json", "data/palettes/*.json"]
